requirements.txt
Flask
gunicorn
PyYAML
requests
